Breast cancer stage prediction decision support mechanism application.
This project is an AI Health Application Project. With the artificial intelligence models in this application, doctors are provided with a decision support mechanism for breast cancer stage prediction.
The desired clinical data is entered by the doctor and the application predicts the stage of the cancer tumor (along with the probability of stage).
During the project, different models were developed with Ada Boost, XG Boost, Rondam Forest and Decision Tree algorithms. These models were evaluated with different measurements such as F1 score, Recall, Precision, Accurcy. Additionally, confusion matrices were created and examined, and the Ada Boost model, which gave the best results, was implemented into the application.
The application was developed using Flask.
An easy-to-use user interface has been designed.
​